量化平台本会话完成卖零股修复、项18监控、P4拆分，三路盲审13个S/A级问题全部修复，397测试全绿

+++ server/src/strategy_framework/adapters.py
@@ -24,7 +24,7 @@ import time
 class Order:
     symbol: str
     action: str  # BUY / SELL
-    volume: int = 0
+    volume: float = 0   # A股=int 股数；crypto=float（BTC 可下 0.001--盲审 A-1 根修，int 截断丢整单）
     price: float = 0.0
     order_type: str = "limit"  # limit / market
     client_id: str = ""
